Add icon for pre-selected attributions, remove gradients
The gradients were hard to understand and their visibility was
not ideal. This adds an icon with hover text instead.

Additionally, we add a unique ID to the package card to avoid
duplicate key errors, which were probably already a problem
before but have become more likely since we now have many more
icons.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>