Fix bug when selecting frequent licenses
There was some hidden magic that converted the license name the user wrote to uppercase letters if there was a match with a frequent license. This logic is removed and the user has to actively choose the frequent license now.
Only in the case of a perfect match (user input: MIT, frequent license: MIT) the frequent license is automatically choosen.
Also removed some formatting, which was confusing for the user and added an integration test.

Signed-off-by: Nico Carl <nico.carl@tngtech.com>
Signed-off-by: Nico <nicocarl@protonmail.com>