Add context menu for PackagePanelCards
This commit adds a context menu for external and manual attributions PackageCards to allow for faster execution of operations on attributions (delete, confirm, hide, show resources). This commit only adds the correctly displayed options and not the operations themself. There are different options, depending on external / manual attribution and whether the attributions belongs to multiple resources.

Signed-off-by: Jonas Tai <jonas.tai@tngtech.com>