Include resources into detailed component list
Context:
In an earlier pull request, the resources were removed to enhance performance. Here, resources are reintroduced, but without the recursive scanning of the tree that was causing the performance issues. To fix this, only the resources, that are directly linked to the respective attributions are included (vs all children that get the attributions from the parents).

Signed-off-by: Anton Bauhofer <anton.bauhofer@tngtech.com>