Fix ContextMenu displayed buttons logic
In views, where no specific resource id is currently related to the attribution (e.g. in the Attribution View), the ContextMenu should only show a "Delete Globally" and "Confirm Globally" buttons and no button to delete the attribution for the related resource. This commit fixes this logic and adds test for the different cases.

Signed-off-by: Jonas Tai <jonas.tai@tngtech.com>