Fix hidden signals in Signals in Folder Content tab
Hiding a signal should remove it from the 'Signals in Folder Content' tab and make it only visible through the 'Signals' tab of the resources, the signal is attached to. So far, the signal stayed greyed out in the 'Signals in Folder Content' tab, if at least one resource, to which the signal was attached to, had one or more other non-hidden signals. This commit fixes this behaviour by adding a filter for resolved attributions to the 'Signals in Folder Content' tab.

Signed-off-by: Jonas Tai <jonas.tai@tngtech.com>