Sort package cards by displayed title
The sorting in the attribution lists only compared the package names of the attribution to sort the attributions. This was confusing for the user, because the card displays additional / different information, depending on the available information (e.g. package name and package version combined, only the comment,...). This commit adds a sorting by the first line that is displayed in the package card with special characters after alphabetical characters. Empty attributions are sorted to the bottom.

Signed-off-by: Jonas Tai <jonas.tai@tngtech.com>