Use a worker for the top progress bar
This commit aims to speed up saving. The following two functions were identified as taking up most of the time when saving a large file:
getUpdatedProgressBarData (~1300ms on my setup)
saveManualAndResolvedAttributionsToFile (~300ms on my setup)
getUpdatedProgressBarData is a function that calculates data for the progress bar at the top of the UI. Since it takes up a lot of time and doesn't need to be blocking, it was moved to a worker where it can be executed asynchronously.

Issue: #1110

Signed-off-by: Markus Obendrauf <markusobendrauf@fb.com>