Attribution Wizard: Add package card context menu and wizard popup skeleton
The Attribution Wizard is meant to streamline the editing of
attributions. Rightclicking a package card (manual attributions only)
and clicking 'Open attribution wizard' opens the popup. The skeleton
comprises a header, the ID (path) of the current resource, and buttons
to cancel and switch to the next view (only a dummy so far).

Fix: #1277
Fix: #1276

Signed-off-by: Lennart Holstein <lennart.holstein@tngtech.com>