Add .opossum support: Wire FileSupportPopup
Based on the user input in the FileSupportPopup, data loading is handled
in different ways. Either converting the outdated format into .opossum
or keeping the old format and proceed. If an outdated file is opened and
a .opossum file with the same name exists, the .opossum file is opened
by default after showing a corresponding popup.

Signed-off-by: Lennart Holstein <lennart.holstein@tngtech.com>