Fix #1868: Fix computation of "files with only external attribution count"
When calculating the number of "Files with only signals" for the tooltip of the progress bar, we want to sum up all files that only have a signal and all files that only have a parent which only has a signal. Before this fix, the logic was such that only files with a direct parent with signal were counted, but not files with a parent higher up the tree, that has a signal.

Signed-off-by: Anton Bauhofer <anton.bauhofer@tngtech.com>