Open locator popup with shortcut
The popup to locate signals can now be opened via shortcut Ctrl+L/Cmd+L or via the top bar if an Opossum file is opened.
This commit also includes a small refactoring which extracts a method to choose the correct icon based on the theme (dark or light) used.

Issue: #1947
Signed-off-by: Meret Behrens <meret.behrens@tngtech.com>