Add license selector to popup
This commit adds an auto-completion input field with a dropdown to the locator popup. The dropdown contains all licenses found in any external attribution and does not aggregate licenses. If a license is selected and the apply button is pressed the corresponding state is set. If the clear button is pressed the state will be reset.
Fixes #1941.

Signed-off-by: Meret Behrens <meret.behrens@tngtech.com>