Update Tooltip text of the criticality bar.
The previous tooltip text was missleading as it was not clear what the
"only" related to, e.g., "Files with only highly critical signals".
Therefore, the tooltip was updated. The new headline makes it clear,
that only resources with signals and without any attributions are
considered.
The logic used to calculate the numbers does not have to be changed, as it
already implements to requested behaviour.

Issue: #1986
Fix: #1986

Signed-off-by: Florian Schepers <florianschepers@gmx.de>