Handle unsaved changes when locating attributions
It is possible to open the ProjectStatisticsPopup or the LocatorPopup
while having unsaved changes of an attribution. If the user then locates
attributions, the UnsavedChangesPopup appears. After the user has decided
to save or undo the changes (or cancelled), the ProjectStatisticsPopup
or LocatorPopup is shown again. From here, the user can again initialize
the location process.

Fix: #2020

Signed-off-by: Lennart Holstein <lennart.holstein@tngtech.com>