Update preferences when a child attribution changes
Update preferences when an attribution is created, deleted, linked or unlinked.

Manual attributions act as break points for preference calculations. Therefore, any time we change an attribution placement in the resource tree, we need to update the preference calculations of its parents. Only the parents are relevant, because the 'preferred over' calculation travels down the resource tree only.

Issue: #2030
Fix: #2030

Signed-off-by: Markus Obendrauf <markusobendrauf@fb.com>