Update tooltip text of the criticality bar for improved understandability
The previous tooltip text caused irritations about whether the first
number 'with signals' is the total number, i.e., the width of the total
bar, or not. The changes introduced in this diff should make clear what
is the total number of resources and what are the respective subparts of resources with highly-,
medium- or non-critical signals.

Signed-off-by: Florian Schepers <florianschepers@gmx.de>