Fix sorting of licenses in project statistics popup
We here fix the sorting of licenses in the "Critical Licenses" and the "Signals per Sources" tables.
In the "Critical Licenses" table, we are showing the entries with "high" criticality alphabetically sorted on top, and those with "medium" criticality alphabetically sorted below.
In the "Signals per Sources" table, on the other hand, we are showing all entries alphabetically sorted by license name.

Fixes #2050

Signed-off-by: Anton Bauhofer <anton.bauhofer@tngtech.com>