Cache data on Folder Progress Bar Worker
Currently, all information needed for calculating the folder progress
bar is sent to the worker on every click on a new resource. That is slow
and not necessary, as most of the data just changes after saving.
Therefore, in this commit, the progress-bar-worker is modified such that
it caches information and only reloads the cache after saving.

Signed-off-by: Florian Schepers <florianschepers@gmx.de>