Add feature to directly open a file via ENV variable
In addition to opening a file from a cli argument this commit adds the function to open a file by setting an environment variable. If the ENV variable OPOSSUM_FILE is set, the app will try to open this file directly when starting. To simplify the usage of this feature and make it independent of the operating system the yarn start command now accepts an optional argument and sets the environment variable when starting the app.

Signed-off-by: Meret Behrens <meret.behrens@tngtech.com>