chore: add initial opossum file to e2e test output artefacts
This is helpful because currently the opossum file is saved in the state in which
the error happened. If one wants to reproduce the steps that let to the error by
hand, that is not possible in all cases (because steps can't necessarily be reverted).

This is especially helpful if a bug in a release is found by the test. One can use
the initial opossum file and go through the test step by step to check if the
failure also happens in the released app.

Signed-off-by: Alexander zur Bonsen <alexander.zur.bonsen@tngtech.com>