fix: don't show original license info if signal is of type first party
in the attribution column we don't show license info fields (name, text,
copyright) if an attribution is first party. Thus, we leave this fields
empty in the diff popup as well. We did so for the current attribution
already, but not for the original signal since the assumption was that it
shouldn't come with such info. But the fields were only removed recently
and we still get first party signals with license info. Thus, we also
remove them from original signals in the diff popup now.

Signed-off-by: alexzurbonsen <alexander.zur.bonsen@tngtech.com>