fix: use 'upstream address' naming convention
Replace the text 'Component URL' with the preferred terminology
'Upstream Address'.

The field in the opossum file and the corresponding variable name in code
were not updated, as this would be a breaking change.

Also, change the open link in browser text from 'Open component URL in
browser' to 'Open in browser' - IMO it's obvious from the context what
we're opening in the browser, and the extra text only makes it take longer
to understand.

Signed-off-by: Markus Obendrauf <markus.obendrauf@tngtech.com>