feat: add url validation logic
Previously, the URL field did not have any validation logic. This PR adds validation logic that checks whether the input
can be parsed as an URL with some best effort adjustments. Note that this does not break opossum files that already contain
bad data. The field will be highlighted but one can still convert a signal to an attribution. Only if the user edits some
field, they will need to correct the bad URL in order to proceed.